+++ Webservice/Webservice.php
@@ -63,6 +63,8 @@ class Webservice
 
     const MAGENTO_STATUS_DISABLE                             = 2;
 
+    const MAGENTO_PRODUCT_UPDATE_USELESS                     = 2;
+
     const ADMIN_STOREVIEW                                    = 0;
 
     protected $client;
@@ -298,7 +300,6 @@ class Webservice
         ) {
             $resource = self::SOAP_ACTION_CATALOG_PRODUCT_CREATE;
         } else {
-            $productPart = $this->removeNonUpdatePart($productPart);
             $resource = self::SOAP_ACTION_CATALOG_PRODUCT_UPDATE;
         }
 
@@ -335,10 +336,14 @@ class Webservice
      */
     public function sendNewCategory(array $category)
     {
-        return $this->client->call(
+        $categoryId =  $this->client->call(
             self::SOAP_ACTION_CATEGORY_CREATE,
             $category
         );
+
+        $this->updateCategoryIfMultipleStoreView($category, $categoryId);
+
+        return $categoryId;
     }
 
     /**
@@ -868,23 +873,6 @@ class Webservice
     }
 
     /**
-     * Cleanup part of the product data that should not be sent as
-     * update part
-     *
-     * @param array $productPart
-     *
-     * @return array
-     */
-    protected function removeNonUpdatePart(array $productPart)
-    {
-        if (isset($productPart[1]['url_key'])) {
-            unset($productPart[1]['url_key']);
-        }
-
-        return $productPart;
-    }
-
-    /**
      * Update product if there is multiple Magento store views.
      *
      * @param array $productPart
@@ -895,13 +883,30 @@ class Webservice
 
         if (count($storeViewList) > 1 && count($productPart) === static::CREATE_PRODUCT_SIZE) {
             $updateProductPart = array_merge(
-                array_slice($productPart, static::MAGENTO_STATUS_DISABLE),
+                array_slice($productPart, static::MAGENTO_PRODUCT_UPDATE_USELESS),
                 ['sku']
             );
             $this->client->addCall([static::SOAP_ACTION_CATALOG_PRODUCT_UPDATE, $updateProductPart]);
 
             $updateProductPart[2] = static::ADMIN_STOREVIEW;
             $this->client->addCall([static::SOAP_ACTION_CATALOG_PRODUCT_UPDATE, $updateProductPart]);
+        }
+    }
+
+    /**
+     * @param array  $category
+     * @param string $categoryId
+     */
+    protected function updateCategoryIfMultipleStoreView($category, $categoryId)
+    {
+        $storeViewList = $this->getStoreViewsList();
+
+        if (count($storeViewList) > 1) {
+            $category[0] = $categoryId;
+            $this->client->addCall([static::SOAP_ACTION_CATEGORY_UPDATE, $category]);
+
+            $category[2] = static::ADMIN_STOREVIEW;
+            $this->client->addCall([static::SOAP_ACTION_CATEGORY_UPDATE, $category]);
         }
     }
 }